+++ resources/js/src/components/mollieContainer/Label.js
@@ -1,0 +1,7 @@
+export const Label = ({ item }) => {
+    return (
+        <>
+            <div dangerouslySetInnerHTML={{ __html: item.label }}/>
+        </>
+    );
+};
